pass inits to mem-sm

+++ state-machines/mem-common/src/mem_module_check_point.rs
@@ -1,27 +1,27 @@
 #[derive(Debug, Default, Clone)]
 pub struct MemModuleCheckPoint {
-    pub from_addr: u32,
+    pub from_addr: u64,
     pub from_skip: u32,
-    pub to_addr: u32,
+    pub to_addr: u64,
     pub to_count: u32,
     pub count: u32,
 }
 
 impl MemModuleCheckPoint {
-    pub fn new(from_addr: u32, skip: u32, count: u32) -> Self {
+    pub fn new(from_addr: u64, skip: u32, count: u32) -> Self {
         Self { from_addr, from_skip: skip, to_addr: from_addr, to_count: count, count }
     }
     pub fn init(
         &mut self,
-        from_addr: u32,
+        from_addr: u64,
         skip: u32,
-        to_addr: u32,
+        to_addr: u64,
         to_count: u32,
         count: u32,
     ) -> Self {
         Self { from_addr, from_skip: skip, to_addr, to_count, count }
     }
-    pub fn add_rows(&mut self, addr: u32, count: u32) {
+    pub fn add_rows(&mut self, addr: u64, count: u32) {
         // data is processed by order address, an only one address by chunk/step
         // TODO: assert -> debug_assert
         assert!(addr >= self.to_addr);
