+++ state-machines/mem/src/mem_inputs.rs
@@ -11,7 +11,7 @@ fn format_u64_hex(value: u64) -> String {
 
 #[derive(Debug, Clone)]
 pub struct MemAlignInput {
-    pub addr: u32,
+    pub addr: u64,
     pub is_write: bool,
     pub width: u8,
     pub step: u64,
@@ -21,7 +21,7 @@ pub struct MemAlignInput {
 
 #[derive(Debug)]
 pub struct MemInput {
-    pub addr: u32,      // address in word native format means byte_address / MEM_BYTES
+    pub addr: u64,      // address in word native format means byte_address / MEM_BYTES
     pub is_write: bool, // it's a write operation
     pub step: u64,      // mem_step = f(main_step, main_step_offset)
     pub value: u64,     // value to read or write
