+++ state-machines/mem/src/mem_module_instance.rs
@@ -17,9 +17,9 @@ pub struct MemModuleInstance<F: PrimeField64> {
     module: Arc<dyn MemModule<F>>,
 
     check_point: MemModuleSegmentCheckPoint,
-    min_addr: u32,
+    min_addr: u64,
     #[allow(dead_code)]
-    max_addr: u32,
+    max_addr: u64,
 }
 
 impl<F: PrimeField64> MemModuleInstance<F> {
