+++ state-machines/mem/src/mem_module_planner.rs
@@ -8,7 +8,7 @@ use zisk_common::{CheckPoint, ChunkId, InstanceType, Plan, SegmentId};
 pub struct MemModulePlanner {
     config: MemModulePlannerConfig,
     rows_available: u32,
-    last_addr: u32, // addr of last addr uses
+    last_addr: u64, // addr of last addr uses
 
     segments: Vec<MemModuleSegmentCheckPoint>,
     current_segment_chunks: HashMap<ChunkId, MemModuleCheckPoint>,
@@ -17,7 +17,7 @@ pub struct MemModulePlanner {
     last_chunk: Option<ChunkId>,
     current_chunk_id: Option<ChunkId>,
     reference_addr_chunk: Option<ChunkId>,
-    reference_addr: u32,
+    reference_addr: u64,
     reference_skip: u32,
     cursor: MemCountersCursor,
 }
@@ -27,7 +27,7 @@ pub struct MemModulePlannerConfig {
     pub airgroup_id: usize,
     pub air_id: usize,
     pub addr_index: usize,
-    pub from_addr: u32,
+    pub from_addr: u64,
     pub rows: u32,
     pub consecutive_addr: bool,
 }
@@ -79,14 +79,14 @@ impl MemModulePlanner {
     /// If the chunk_id is not in the list, it will be added. This method need to verify the
     /// distance between the last addr-step and the current addr-step, if the distance is
     /// greater than MEMORY_MAX_DIFF we need to add extra intermediate "steps".
-    fn add_to_current_instance(&mut self, chunk_id: ChunkId, addr: u32, count: u32) {
+    fn add_to_current_instance(&mut self, chunk_id: ChunkId, addr: u64, count: u32) {
         self.set_current_chunk_id(chunk_id);
         let intermediate_rows = self.add_intermediates(addr);
         self.preopen_segment(addr, intermediate_rows);
         self.set_reference(chunk_id, addr);
         self.add_rows(addr, count);
     }
-    fn set_reference(&mut self, chunk_id: ChunkId, addr: u32) {
+    fn set_reference(&mut self, chunk_id: ChunkId, addr: u64) {
         self.reference_addr_chunk = Some(chunk_id);
         self.reference_addr = addr;
         self.reference_skip = 0;
@@ -122,12 +122,12 @@ impl MemModulePlanner {
         // all rows are available
         self.rows_available = self.config.rows;
     }
-    fn add_next_addr_to_segment(&mut self, addr: u32) {
+    fn add_next_addr_to_segment(&mut self, addr: u64) {
         let chunk_id = self.current_chunk_id.unwrap();
         self.add_chunk_to_segment(chunk_id, addr, 1, 0);
     }
 
-    fn add_chunk_to_segment(&mut self, chunk_id: ChunkId, addr: u32, count: u32, skip: u32) {
+    fn add_chunk_to_segment(&mut self, chunk_id: ChunkId, addr: u64, count: u32, skip: u32) {
         if self.current_segment_chunks.is_empty() {
             self.first_chunk_id = Some(chunk_id);
         }
@@ -136,7 +136,7 @@ impl MemModulePlanner {
             .and_modify(|checkpoint| checkpoint.add_rows(addr, count))
             .or_insert(MemModuleCheckPoint::new(addr, skip, count));
     }
-    fn preopen_segment(&mut self, addr: u32, intermediate_rows: u32) {
+    fn preopen_segment(&mut self, addr: u64, intermediate_rows: u32) {
         if self.rows_available == 0 {
             if intermediate_rows > 0 {
                 // prevent last intermediate row zero
@@ -145,7 +145,7 @@ impl MemModulePlanner {
             self.open_segment();
         }
     }
-    fn consume_rows(&mut self, addr: u32, rows: u32, skip: u32) {
+    fn consume_rows(&mut self, addr: u64, rows: u32, skip: u32) {
         if rows == 0 && self.rows_available > 0 {
             return;
         }
@@ -164,7 +164,7 @@ impl MemModulePlanner {
         self.rows_available -= rows;
         self.reference_skip += rows;
     }
-    fn consume_intermediate_rows(&mut self, addr: u32, rows: u32, skip: u32) {
+    fn consume_intermediate_rows(&mut self, addr: u64, rows: u32, skip: u32) {
         if rows == 0 && self.rows_available > 0 {
             return;
         }
@@ -182,7 +182,7 @@ impl MemModulePlanner {
         self.rows_available -= rows;
     }
 
-    fn add_intermediate_rows(&mut self, addr: u32, count: u32) {
+    fn add_intermediate_rows(&mut self, addr: u64, count: u32) {
         let mut pending = count;
 
         while pending > 0 {
@@ -193,7 +193,7 @@ impl MemModulePlanner {
         }
     }
 
-    fn add_rows(&mut self, addr: u32, count: u32) {
+    fn add_rows(&mut self, addr: u64, count: u32) {
         let mut pending = count;
         while pending > 0 {
             let rows = min(pending, self.rows_available);
@@ -202,18 +202,18 @@ impl MemModulePlanner {
             pending -= rows;
         }
     }
-    fn add_intermediate_addr(&mut self, from_addr: u32, to_addr: u32) {
+    fn add_intermediate_addr(&mut self, from_addr: u64, to_addr: u64) {
         // adding internal reads of zero for consecutive addresses
         let count = to_addr - from_addr + 1;
         if count > 1 {
             self.add_intermediate_rows(from_addr, 1);
-            self.add_intermediate_rows(to_addr, count - 1);
+            self.add_intermediate_rows(to_addr, (count - 1) as u32);
         } else {
             assert_eq!(to_addr, from_addr);
             self.add_intermediate_rows(to_addr, 1);
         }
     }
-    fn add_intermediates(&mut self, addr: u32) -> u32 {
+    fn add_intermediates(&mut self, addr: u64) -> u32 {
         if self.last_addr != addr {
             if self.config.consecutive_addr && (addr - self.last_addr) > 1 {
                 self.add_intermediate_addr(self.last_addr + 1, addr - 1);
